add snackbar in all sections that require internet connection
