+++ scp/android/gradle.properties
@@ -2,3 +2,4 @@ org.gradle.jvmargs=-Xmx1536M
 
 android.enableJetifier=true
 android.useAndroidX=true
+android.enableR8=true
